tools: rearrange source files
Consolidate property reading into a single source file.
