--- tools/openslide-show-properties.c
@@ -1,73 +1,0 @@
-/*
- *  OpenSlide, a library for reading whole slide image files
- *
- *  Copyright (c) 2007-2012 Carnegie Mellon University
- *  All rights reserved.
- *
- *  OpenSlide is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Lesser General Public License as
- *  published by the Free Software Foundation, version 2.1.
- *
- *  OpenSlide is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public
- *  License along with OpenSlide. If not, see
- *  <http://www.gnu.org/licenses/>.
- *
- */
-
-#include <stdbool.h>
-#include <stdio.h>
-#include <glib.h>
-#include "openslide.h"
-#include "openslide-common.h"
-#include "slidetool.h"
-
-static bool process(const char *file, int successes, int total) {
-  g_autoptr(openslide_t) osr = openslide_open(file);
-  if (common_warn_on_error(osr, "%s", file)) {
-    return false;
-  }
-
-  // print header
-  if (successes > 0) {
-    printf("\n");
-  }
-  if (total > 1) {
-    // format inspired by head(1)/tail(1)
-    printf("==> %s <==\n", file);
-  }
-
-  // read properties
-  const char * const *property_names = openslide_get_property_names(osr);
-  while (*property_names) {
-    const char *name = *property_names;
-    const char *value = openslide_get_property_value(osr, name);
-    printf("%s: '%s'\n", name, value);
-
-    property_names++;
-  }
-
-  return true;
-}
-
-static int do_show_properties(int narg, char **args) {
-  int successes = 0;
-  for (int i = 0; i < narg; i++) {
-    if (process(args[i], successes, narg)) {
-      successes++;
-    }
-  }
-  return successes != narg;
-}
-
-const struct command show_properties_cmd = {
-  .parameter_string = "FILE...",
-  .summary = "Print OpenSlide properties for a slide.",
-  .options = legacy_opts,
-  .min_positional = 1,
-  .handler = do_show_properties,
-};
