+++ tools/slidetool-image.c
