+++ tools/slidetool-prop.c
@@ -1,7 +1,7 @@
 /*
  *  OpenSlide, a library for reading whole slide image files
  *
- *  Copyright (c) 2010-2012 Carnegie Mellon University
+ *  Copyright (c) 2007-2012 Carnegie Mellon University
  *  All rights reserved.
  *
  *  OpenSlide is free software: you can redistribute it and/or modify
@@ -26,7 +26,45 @@
 #include "openslide-common.h"
 #include "slidetool.h"
 
-static bool process(const char *file) {
+static bool list_props(const char *file, int successes, int total) {
+  g_autoptr(openslide_t) osr = openslide_open(file);
+  if (common_warn_on_error(osr, "%s", file)) {
+    return false;
+  }
+
+  // print header
+  if (successes > 0) {
+    printf("\n");
+  }
+  if (total > 1) {
+    // format inspired by head(1)/tail(1)
+    printf("==> %s <==\n", file);
+  }
+
+  // read properties
+  const char * const *property_names = openslide_get_property_names(osr);
+  while (*property_names) {
+    const char *name = *property_names;
+    const char *value = openslide_get_property_value(osr, name);
+    printf("%s: '%s'\n", name, value);
+
+    property_names++;
+  }
+
+  return true;
+}
+
+static int do_show_properties(int narg, char **args) {
+  int successes = 0;
+  for (int i = 0; i < narg; i++) {
+    if (list_props(args[i], successes, narg)) {
+      successes++;
+    }
+  }
+  return successes != narg;
+}
+
+static bool quickhash1sum(const char *file) {
   g_autoptr(openslide_t) osr = openslide_open(file);
   if (common_warn_on_error(osr, "%s", file)) {
     return false;
@@ -46,12 +84,20 @@ static bool process(const char *file) {
 static int do_quickhash1sum(int narg, char **args) {
   int ret = 0;
   for (int i = 0; i < narg; i++) {
-    if (!process(args[i])) {
+    if (!quickhash1sum(args[i])) {
       ret = 1;
     }
   }
   return ret;
 }
+
+const struct command show_properties_cmd = {
+  .parameter_string = "FILE...",
+  .summary = "Print OpenSlide properties for a slide.",
+  .options = legacy_opts,
+  .min_positional = 1,
+  .handler = do_show_properties,
+};
 
 const struct command quickhash1sum_cmd = {
   .parameter_string = "FILE...",
